refactoring board.js
- board.js를 refactoring
