+++ my-app/src/square.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-export  default class Square extends React.Component {
+export class Square extends React.Component {
     render() {
       return (
       <button className="square" onClick={() => this.props.onClick()}>
